Restructure BPMN, DMN, and form definitions
Rename the 'bpmn' directory into 'models' to reduce the risk of
confusion with parts of the library. Keep the camunda examples separate
as before, but move everything from 'tutorial' one level up.
Create a separate directory for JSON-form definitions in the repository
root. Modify the readme file and the curses UI accordingly.